feat: added random country logics
