--- tests/testsData/requestBodies/mayotte-polygon-variable.json
@@ -1,183 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [44.7436675, -12.6615483],
-            [44.7436856, -12.6590777],
-            [44.7437391, -12.6551207],
-            [44.7447262, -12.6378938],
-            [44.7476921, -12.6183879],
-            [44.7525899, -12.5992613],
-            [44.7544333, -12.5936132],
-            [44.7555382, -12.590392],
-            [44.7604767, -12.5774769],
-            [44.7690772, -12.5596558],
-            [44.7794132, -12.5427481],
-            [44.7865404, -12.5329373],
-            [44.7887214, -12.5301177],
-            [44.7935659, -12.5240968],
-            [44.8070578, -12.5094935],
-            [44.82194, -12.496259],
-            [44.8380694, -12.4845204],
-            [44.8552907, -12.4743908],
-            [44.8734382, -12.4659675],
-            [44.8912508, -12.4596581],
-            [44.9086182, -12.454384],
-            [44.9236158, -12.4396371],
-            [44.9236989, -12.4395555],
-            [44.9385712, -12.4263144],
-            [44.9546911, -12.4145688],
-            [44.9719033, -12.4044316],
-            [44.9900422, -12.3960003],
-            [45.0089335, -12.389356],
-            [45.0283957, -12.3845624],
-            [45.0389132, -12.3827837],
-            [45.0505713, -12.3811183],
-            [45.0598996, -12.3799999],
-            [45.0799383, -12.379027],
-            [45.0915505, -12.3793478],
-            [45.106539, -12.3801805],
-            [45.1149662, -12.3808204],
-            [45.1348138, -12.3837052],
-            [45.1542788, -12.3884871],
-            [45.1731742, -12.39512],
-            [45.1913182, -12.4035403],
-            [45.2035783, -12.4105168],
-            [45.2577228, -12.443829],
-            [45.2626823, -12.4469783],
-            [45.2694414, -12.4518945],
-            [45.2820764, -12.4585643],
-            [45.2987442, -12.4683919],
-            [45.3148786, -12.4801221],
-            [45.3277536, -12.491403],
-            [45.3445052, -12.5073736],
-            [45.3465185, -12.5093195],
-            [45.3542294, -12.5173009],
-            [45.3625852, -12.5264378],
-            [45.3641915, -12.5277674],
-            [45.3768354, -12.5391565],
-            [45.3803356, -12.5429396],
-            [45.382194, -12.5442895],
-            [45.3970891, -12.5575123],
-            [45.4043932, -12.5651613],
-            [45.4094745, -12.5689881],
-            [45.4149828, -12.5735763],
-            [45.4253946, -12.582563],
-            [45.4347849, -12.5911957],
-            [45.4482945, -12.6057863],
-            [45.4602845, -12.6216067],
-            [45.4706392, -12.6385045],
-            [45.4738994, -12.6447259],
-            [45.4776257, -12.6521511],
-            [45.4829854, -12.6637424],
-            [45.4897871, -12.6822988],
-            [45.4947046, -12.7014202],
-            [45.4959316, -12.7079784],
-            [45.4975953, -12.7177829],
-            [45.4976387, -12.7181021],
-            [45.4977453, -12.7186794],
-            [45.4985681, -12.723667],
-            [45.5022704, -12.7402191],
-            [45.5030229, -12.7437327],
-            [45.5060096, -12.7632351],
-            [45.5070347, -12.7829308],
-            [45.5069015, -12.7905808],
-            [45.5065767, -12.7994047],
-            [45.5057632, -12.8114541],
-            [45.5028531, -12.8309677],
-            [45.4980075, -12.8501072],
-            [45.4949617, -12.8585104],
-            [45.4941528, -12.8708094],
-            [45.4912414, -12.8903228],
-            [45.4863938, -12.9094621],
-            [45.48222, -12.9216302],
-            [45.476111, -12.9377179],
-            [45.4735472, -12.9441307],
-            [45.4649836, -12.9619739],
-            [45.4546766, -12.9789076],
-            [45.4458336, -12.9906437],
-            [45.4466078, -13.0055616],
-            [45.4456562, -13.0252607],
-            [45.4453241, -13.0283736],
-            [45.4443427, -13.0369292],
-            [45.441758, -13.0533291],
-            [45.436904, -13.0724675],
-            [45.4301589, -13.0910472],
-            [45.4251252, -13.1020649],
-            [45.4208383, -13.110739],
-            [45.4173002, -13.1175632],
-            [45.4069838, -13.134495],
-            [45.3950223, -13.1503539],
-            [45.3815308, -13.1649867],
-            [45.366639, -13.1782523],
-            [45.3512286, -13.1895358],
-            [45.3246854, -13.2071279],
-            [45.3007291, -13.2234023],
-            [45.2980208, -13.2252107],
-            [45.2807664, -13.235369],
-            [45.2669312, -13.2419902],
-            [45.2406801, -13.2533758],
-            [45.236326, -13.2552054],
-            [45.2173763, -13.2618661],
-            [45.1978504, -13.2666739],
-            [45.1834096, -13.2689777],
-            [45.1616434, -13.2716721],
-            [45.1561706, -13.2722763],
-            [45.1360614, -13.2732554],
-            [45.1200359, -13.2726495],
-            [45.0827758, -13.269805],
-            [45.0786912, -13.2694526],
-            [45.0587745, -13.2665642],
-            [45.0392436, -13.2617761],
-            [45.0297122, -13.258694],
-            [45.0030928, -13.2493705],
-            [44.9936681, -13.2458099],
-            [44.97547, -13.2373768],
-            [44.9695202, -13.234135],
-            [44.9517683, -13.2240798],
-            [44.9404536, -13.2171797],
-            [44.939193, -13.2163473],
-            [44.9155735, -13.2006256],
-            [44.9006715, -13.1897041],
-            [44.8964941, -13.1862498],
-            [44.8728998, -13.1662026],
-            [44.8621719, -13.1564057],
-            [44.849889, -13.1432337],
-            [44.8367357, -13.1278027],
-            [44.835513, -13.1263544],
-            [44.8235376, -13.1105059],
-            [44.8132062, -13.0935829],
-            [44.809811, -13.0870618],
-            [44.8033011, -13.0739742],
-            [44.7981088, -13.0626607],
-            [44.7972518, -13.0605952],
-            [44.7897635, -13.0422566],
-            [44.7882582, -13.0380474],
-            [44.7851886, -13.0298933],
-            [44.7560831, -12.9423726],
-            [44.7547554, -12.938227],
-            [44.749889, -12.9190917],
-            [44.7469585, -12.8995805],
-            [44.7459919, -12.8798815],
-            [44.746998, -12.8601845],
-            [44.7499667, -12.8406788],
-            [44.7520765, -12.8314567],
-            [44.770136, -12.7603591],
-            [44.7677481, -12.7563061],
-            [44.7591744, -12.73847],
-            [44.752424, -12.7198946],
-            [44.7475616, -12.700759],
-            [44.7446335, -12.6812475],
-            [44.7436676, -12.6615483]
-          ]
-        ]
-      }
-    }
-  ]
-}
