--- tests/testsData/requestBodies/sicily-polygon.json
@@ -1,30 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {},
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [15.778648658930864, 39.76616859292575],
-            [15.77373976387289, 39.300461098551686],
-            [15.783312109235899, 39.30236041046368],
-            [15.491151038370289, 38.765991042512645],
-            [14.641298581463026, 38.24686206662315],
-            [13.42790153805581, 38.27984870569096],
-            [12.380956944571746, 38.19531256495158],
-            [12.08049165956653, 37.64280441340706],
-            [12.930466838850148, 37.17665885497673],
-            [14.244455323486596, 36.65913244429737],
-            [15.369164996180528, 36.70981651725083],
-            [16.181955295400105, 37.84207037693634],
-            [17.683054496662884, 39.45347501981596],
-            [15.778648658930864, 39.76616859292575]
-          ]
-        ]
-      }
-    }
-  ]
-}
